more fixes; added new service with new pages

+++ src/main/java/com/UCH/UAContentHub/Repository/LikesRepository.java
@@ -4,6 +4,6 @@ package com.UCH.UAContentHub.Repository;
 import com.UCH.UAContentHub.Entity.Likes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LikesRepository extends JpaRepository<Likes,Long> {
+public interface LikesRepository extends JpaRepository<Likes,Integer> {
 
 }
