+++ src/main/java/com/UCH/UAContentHub/Repository/Profile_has_tagsRepository.java
@@ -4,5 +4,5 @@ package com.UCH.UAContentHub.Repository;
 import com.UCH.UAContentHub.Entity.Profile_has_tags;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Profile_has_tagsRepository extends JpaRepository<Profile_has_tags,Long> {
+public interface Profile_has_tagsRepository extends JpaRepository<Profile_has_tags,Integer> {
 }
