+++ src/main/java/com/UCH/UAContentHub/Repository/СomplaintRepository.java
@@ -4,6 +4,6 @@ package com.UCH.UAContentHub.Repository;
 import com.UCH.UAContentHub.Entity.Complaint;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface СomplaintRepository extends JpaRepository<Complaint,Long> {
+public interface СomplaintRepository extends JpaRepository<Complaint,Integer> {
 
 }
